// src/lib/api-zod/api.ts
export const UpdateProductBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  prerequisites: zod.string().optional(),
  rewardStructure: zod.enum(["per_subscription", "per_lead"]).optional(),
  subscriptionPrice: zod.number().optional(),
  commissionRate: zod.number().optional(),
  leadBounty: zod.number().nullish(),
  platformFeeRate: zod.number().optional(),
});